requirements.txt
pre-commit==4.2.*
ipywidgets==8.1.*
jsonargparse[signatures]==4.38.*
matplotlib==3.10.*
torchinfo==1.8.*
datasets==3.6.*
tensorboard==2.19.*
